chore(frontend): reorganize postcss config files
