+++ frontend/config/postcss/text-content.config.js
@@ -12,7 +12,8 @@ export default {
           return selector.replace(/^(html|body)/, prefix)
         }
         return prefixedSelector
-      }
-    })
-  ]
+      },
+    }),
+  ],
 }
+
