+++ frontend/config/postcss/xwiki-sandbox-prefixer-options.js
@@ -12,3 +12,4 @@ export const xwikiSandboxPrefixerOptions = {
 }
 
 export default xwikiSandboxPrefixerOptions
+
